+++ packages/models/src/box/sensorLayouts/sensorDefinitions/windspeed.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  title: 'Windgeschwindigkeit',
+  unit: 'm/s',
+  sensorType: 'WINDSPEED',
+  icon: 'osem-particulate-matter'
+};